Realtime_Stock_Market_Prediction-fixed/backend/requirements.txt
fastapi
uvicorn[standard]
yfinance>=0.2.40
numpy
pandas
scikit-learn
xgboost
nltk
scipy
joblib
requests
pymongo
PyJWT
bcrypt
